pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "ccxt>=4.5.7",
    "fpdf>=1.7.2",
    "joblib>=1.5.2",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.50.0",
]
